Regtest integration
Making it possible to run chromanode in regtest mode. Routing all
imports of bitcore through a patch that adds regtest support, see
https://github.com/bitpay/bitcore/issues/1190
Adding a sub-config option to testnet called regtest that sets the
server in regtest mode if both testnet is selected and regtest is
selected in the config file.